+++ binnacle/lib/model/phone/Boat.dart
@@ -9,7 +9,6 @@ import 'dart:math';
 /// the live sailboat and a mocked "ideal" sailboat (generated from algorithm).
 class PhoneBoat extends Boat {
   Geolocator geolocator;
-  AccelerometerEvent test;
   
   PhoneBoat() {
     geolocator = Geolocator();
@@ -28,21 +27,26 @@ class PhoneBoat extends Boat {
 
   void initListAngleStream() {
     listAngle = new StreamController<double>();
-    listAngle.addStream(accelerometerToZRotate(accelerometerEvents));
+    listAngle.addStream(accelerometerToListAngle(accelerometerEvents));
   }
 
   // Converting acclerometer stream into a listAngle stream
   // listAngle/phoneRoll
-  Stream<double> accelerometerToZRotate(Stream<AccelerometerEvent> sensorStream) async* {
-    double zRotation;
+  Stream<double> accelerometerToListAngle(Stream<AccelerometerEvent> sensorStream) async* {
+    double listAngle;
+    // Every event in the acclerometer stream convert value to list angle
     await for (var event in sensorStream) {
       // Conversion for getting the z rotation when potrait
       // from the accelerometer which will be the 
       // phone's way of calculating the
       // angle of list
-      zRotation =  atan2(event.x, event.y);
-      yield zRotation;
+      listAngle =  eventToListAngle(event);
+      yield listAngle;
     }
+  }
+
+  double eventToListAngle(AccelerometerEvent event) {
+    return atan2(event.x, event.y);
   }
 
 }
